indentation: new module for indent and dedent
The main lib.rs file was getting very big and these functions don't
really match the other wrapping functions.
